requirements.txt
altair==5.5.0
altgraph @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
future @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-search-results==2.4.2
googleapis-common-protos==1.69.0
greenlet==3.1.1
gspread==6.2.0
gspread-dataframe==4.0.0
gspread-formatting==1.2.0
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
macholib @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
narwhals==1.36.0
networkx==3.2.1
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.71.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.50.0
preshed==3.0.9
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.29.0
sentence-transformers==4.0.2
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.1
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
